requirements.txt
transformers
torch
pillow
gradio
transformers
torch
pillow
gradio
hf_xet
datasets
accelerate
scikit-learn
numpy
tqdm
evaluate
tensorboard
matplotlib
seaborn
